flask_back_end/requirements.txt
Flask
bson
flask_pymongo
pyjwt
flask_uploads
